+++ create_release.sh
@@ -114,8 +114,8 @@ update_release_body_and_changelog () {
 	sed -i -E "s=(\s)\[([0-9a-f]{8})([0-9a-f]{32})\]= [\2](${GIT_REPO_URL}/commit/\2\3)=g" ./CHANGELOG.md
 
 	# Update changelog to add links to closed issues - comma included!
-	# "closes [#1]," -> "closes [#1](https:/www.../issues/1),""
-	sed -i -r -E "s=closes \[#([0-9]+)\],=closes [#\1](${GIT_REPO_URL}/issues/\1),=g" ./CHANGELOG.md
+	# "closes #1" -> "closes [#1](https:/www.../issues/1)""
+	sed -i -r -E "s=closes \#([0-9]+)=closes [#\1](${GIT_REPO_URL}/issues/\1)=g" ./CHANGELOG.md
 }
 
 # update version in cargo.toml + docker-compose.yml, to match selected current version
@@ -167,8 +167,7 @@ check_tag () {
 				PATCH=$((PATCH + 1))
 				break;;
 			*)
-				error_close "invalid option $REPLY"
-				break;;
+				error_close "invalid option $REPLY";;
 		esac
 	done
 }
@@ -273,8 +272,7 @@ main() {
 	do
 		case $choice in
 			0)
-				exit
-				break;;
+				exit;;
 			1)
 				cargo_build_all
 				main