+++ src/main.rs
@@ -27,27 +27,21 @@ use std::sync::{
     Arc,
 };
 use tokio::sync::broadcast;
-use tracing::Level;
 use word_art::Intro;
 use ws::open_connection;
 
 fn close_signal(light_status: Arc<AtomicBool>) {
     simple_signal::set_handler(&[Signal::Int, Signal::Term], move |_| {
-        light_status.store(false, Ordering::SeqCst);
+        light_status.store(false, Ordering::Relaxed);
         std::thread::sleep(std::time::Duration::from_millis(250));
         std::process::exit(1);
     });
 }
 
 fn setup_tracing(app_envs: &AppEnv) {
-    let level = if app_envs.trace {
-        Level::TRACE
-    } else if app_envs.debug {
-        Level::DEBUG
-    } else {
-        Level::INFO
-    };
-    tracing_subscriber::fmt().with_max_level(level).init();
+    tracing_subscriber::fmt()
+        .with_max_level(app_envs.log_level)
+        .init();
 }
 
 #[tokio::main]