+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# <a href='https://github.com/mrjackwills/belugasnooze_pi/releases/tag/v0.2.1'>v0.2.1</a>
+### 2023-01-06
+
+### Chores
++ dependencies updated, [3036cf7a](https://github.com/mrjackwills/belugasnooze_pi/commit/3036cf7a50168d7df9a696cedf7d88843451755f), [11bd312c](https://github.com/mrjackwills/belugasnooze_pi/commit/11bd312c08d3310454f145c16bfccfcdc2d9848d)
+
+### Docs
++ comments & typos, [8c0235aa](https://github.com/mrjackwills/belugasnooze_pi/commit/8c0235aac0d72e1a03dc9c0c144093ccf37506cc)
+
+### Features
++ log tests, [aec755e0](https://github.com/mrjackwills/belugasnooze_pi/commit/aec755e0fbcdaeda8a2676f66367860dd12408c9)
++ tracing log level into app_env, [740ef13f](https://github.com/mrjackwills/belugasnooze_pi/commit/740ef13f80c8a5ab496474b570235013f0df4055)
+
+### Refactors
++ Dockerfile tweak, [b6b4603e](https://github.com/mrjackwills/belugasnooze_pi/commit/b6b4603e5dd85d41ac02316e53376f6df745b677)
++ lightstatus use relaxed ordering, [58100be2](https://github.com/mrjackwills/belugasnooze_pi/commit/58100be227b48bcc80bcf9e8825c6421ed7b7032)
++ is_connected removed, [9446df66](https://github.com/mrjackwills/belugasnooze_pi/commit/9446df66bc7c734a629973c78c97d9efccffc963)
+
 # <a href='https://github.com/mrjackwills/belugasnooze_pi/releases/tag/v0.2.0'>v0.2.0</a>
 ### 2022-12-15